+++ app/src/main/java/com/tydeya/familycircle/domain/kitchenorganizer/utils/KitchenOrganizerUtils.kt
@@ -5,6 +5,7 @@ import com.tydeya.familycircle.data.constants.Firebase
 import com.tydeya.familycircle.data.kitchenorganizer.food.Food
 import com.tydeya.familycircle.data.kitchenorganizer.food.FoodStatus
 import com.tydeya.familycircle.data.kitchenorganizer.food.MeasureType
+import java.math.BigDecimal
 
 
 fun convertServerDataToFood(documentSnapshot: DocumentSnapshot): Food {
@@ -12,8 +13,8 @@ fun convertServerDataToFood(documentSnapshot: DocumentSnapshot): Food {
             documentSnapshot.get(Firebase.FIRESTORE_FOOD_TITLE).toString(),
             FoodStatus.values()[(documentSnapshot.getLong(Firebase.FIRESTORE_FOOD_STATUS)
                     ?: 0).toInt()],
-            documentSnapshot
-                    .getDouble(Firebase.FIRESTORE_FOOD_QUANTITY_OF_MEASURE) ?: .0,
+            BigDecimal.valueOf(documentSnapshot
+                    .getDouble(Firebase.FIRESTORE_FOOD_QUANTITY_OF_MEASURE) ?: .0),
             MeasureType
                     .values()[(documentSnapshot.getLong(Firebase.FIRESTORE_FOOD_MEASURE_TYPE) ?: 0)
                     .toInt()]
